Added a function to lower case the keys of a json
